+++ Common/includes/lattice_2d.h
@@ -19,11 +19,11 @@ public:
 	[[nodiscard]] double magnetization() const override;
 	[[nodiscard]] double magnetization_diff(size_t i) const override;
 
-	static double magnetization_diff(short old_spin);
+	static double magnetization_diff(int8_t old_spin);
 
 private:
 	const size_t lattice_length;
-	std::vector<short> spins;
+	std::vector<int8_t> spins;
 };
 
 #endif //LATTICE_2D_H
